accessible/tests/mochitest/test_table_struct_tree.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Table accessible tree and table interface tests for XUL trees">

  <script type="application/javascript" 
          src="chrome://mochikit/content/MochiKit/packed.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/treeview.js" />

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/table.js" />

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Test

    var gTree = null;
    var gTreeBox = null;
    var gTreeView = null;

    // gA11yEventDumpID = "debug";

    function doTest()
    {
      // Initialize the tree
      gTree = document.getElementById("table");
      gTreeBox = gTree.treeBoxObject;
      gView = new nsTableTreeView(3);

      gTree.addEventListener("TreeViewChanged", continueTest, false);
      gTreeBox.view = gView;
    }

    function continueTest()
    {
      gTree.removeEventListener("TreeViewChanged", continueTest, false);

      var cellsArray = [
        [kDataCell, kDataCell],
        [kDataCell, kDataCell],
        [kDataCell, kDataCell]
      ];

      testTableStruct(gTree, cellsArray, kTreeColumnHeader);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=512424"
         title="implement IAccessibleTable2">
        Mozilla Bug 512424
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox id="debug"/>
    <tree id="table" flex="1">
      <treecols>
        <treecol id="col" flex="1" label="column"/>
        <treecol id="scol" flex="1" label="column 2"/>
      </treecols>
      <treechildren id="treechildren"/>
    </tree>
  </hbox>

</window>
